+++ SpringMvc/src/main/java/com/victoria/model/Order.java
@@ -4,31 +4,31 @@ import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class Order {
-    private long orderId;
-    private long userId;
+    private BigInteger orderId;
+    private BigInteger userId;
     private BigDecimal orderCost;
     private String status;
-    private long courierId;
-    public Order(long orderId, long userId, BigDecimal orderCost, String status, long courierId) {
+    private BigInteger courierId;
+    public Order(BigInteger orderId, BigInteger userId, BigDecimal orderCost, String status, BigInteger courierId) {
         this(orderId, userId, orderCost, status);
         this.courierId = courierId;
     }
 
-    public Order(long orderId, long userId, BigDecimal orderCost, String status) {
+    public Order(BigInteger orderId, BigInteger userId, BigDecimal orderCost, String status) {
         this.orderId = orderId;
         this.userId = userId;
         this.orderCost = orderCost;
         this.status = status;
     }
 
-    public long getOrderId() {
+    public BigInteger getOrderId() {
         return orderId;
     }
 
-    public long getCourierId() {
+    public BigInteger getCourierId() {
         return courierId;
     }
-    public long getUserId() {
+    public BigInteger getUserId() {
         return userId;
     }
 
@@ -40,15 +40,15 @@ public class Order {
         return status;
     }
 
-    public void setOrderId(long orderId) {
+    public void setOrderId(BigInteger orderId) {
         this.orderId = orderId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(BigInteger userId) {
         this.userId = userId;
     }
 
-    public void setCourierId(long userId) {
+    public void setCourierId(BigInteger userId) {
         this.courierId = userId;
     }
     public void setOrderCost(BigDecimal orderCost) {
