+++ SpringMvc/src/main/java/com/victoria/repository/ItemRepository.java
@@ -2,10 +2,11 @@ package com.victoria.repository;
 
 import com.victoria.model.Item;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ItemRepository {
     List<Item> getAllItems();
-    Item getItemById(long itemId);
+    Item getItemById(BigInteger itemId);
 
 }
