+++ SpringMvc/src/main/java/com/victoria/service/ItemService.java
@@ -2,9 +2,10 @@ package com.victoria.service;
 
 import com.victoria.model.Item;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ItemService {
     List<Item> getAllItems();
-    Item getItemById(long itemId);
+    Item getItemById(BigInteger itemId);
 }
